+++ src/main/java/shcm/shsupercm/fabric/citresewn/mixin/AtlasLoaderMixin.java
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import shcm.shsupercm.fabric.citresewn.pack.PackParser;
+import shcm.shsupercm.fabric.citresewn.cit.resource.PackParser;
 
 import java.util.List;
 import java.util.Map;
